convert examples to use tonc
So we can throw away the bespoke linkscript/crt0 setup and make it much more
obvious what a consumer actually needs to do to make this setup work for them.

The irq handler setup isn't partially hidden anymore in crt0.s, but instead is
all confined to the sole example C file.