fix bug in converting to signed where the src pointer wasnt advanced
in conv2aas, in MISC_ConvSoundToSigned(), the unsigned wav samples are converted
to unsigned. It turns out that the source data pointer was never traversed.

As this completely breaks the use of wav files, I can only speculate that this
might be because perhaps the signed char data pointer cast was introduced only
at a very late stage, and the author forgot to increment the source
position. I'm not sure why the signed cast was introduced in the first place, as
it shouldn't have an effect on code generation. So I took it away. Now that src
and destination are accessed using the same pointer, we get source increments
for free (again?).